requirements-fixed.txt
# Core FastAPI and Web Framework
fastapi
uvicorn
pydantic
python-multipart

# AI/LLM Providers
groq
openai
anthropic
google-generativeai

# Database
asyncpg
psycopg2-binary
sqlalchemy

# Security & Authentication
bcrypt
cryptography
python-jose[cryptography]
passlib[bcrypt]
bleach

# Caching & Storage
redis==7.0.1
diskcache

# Logging
structlog

# CLI & UI
click
rich

# Version Control
gitpython

# Code Analysis & Graph
networkx

# Container & Deployment
docker==7.1.0

# Rate Limiting
slowapi==0.1.9

# Browser Automation
playwright==1.55.0

# Utilities
requests
python-dotenv
tenacity

# Optional: Scheduling (if using timed_automations)
apscheduler

# Optional: Messaging Integrations
# slack-bolt
# python-telegram-bot
